Update README with note about no working directory
If you run the commands listed in usage

```bash
  mkdir repo-git # or whatever
  cd repo-git
  git init
  hg-fast-export.sh -r <repo>
```

you are not given a working directory to start working in. I was caught off-guard by this when I ran `git status` and everything in the repo was listed as deleted. A quick google search indicates I'm not the only one who was surprised.

The wording of the note is a bit awkward, would appreciate suggestions