Fix processing of the --force option to continue to process remaining options
This is a fix to issue #57 - when processing a command similar to

	hg-fast-export --force --quiet -r /path/to/hg/repo

the script stops processing after --force due to a break statement.  This change
also makes the while loop of the command line procesing code clearer by removing the
case syntax in favor of a test on the arg count.