Restore support for mercurial<4.6
e200cec39fac8a8ebb1c7d7690c00d5cb15f0c30 broke mercurial<4.6 support,
which was a Bad Thing. Linuxy types particularly are used to being able
to use the mercurial package from their OS package manager, rather than
needing to mess around with installing extra packages from pip (which is
normally a disaster if you’re not a Python programmer and used to
messing around with virtualenvs or similar), and almost none of those
have Mercurial 4.6 yet. (Ubuntu, for example, is on 4.5.3 in 18.04 LTS,
and 3.7.3 in 16.04 LTS. So probably having 4.6 or later is about two and
a half to four years off!)

The differences in Mercurial 4.6 are small enough that restoring support
is not difficult, and it makes it work for everyone, so it’s worth
doing.

Fixes #132.